Split resolver struct enum variant absence tests
